services/auth-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0

bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
